RF: Separate out inference CLI for faster response
on actions which don't need to import the whole dependency stack,
such as --help and --verbose
